requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
sqlmodel==0.0.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
alembic==1.12.0
pydantic==2.4.2
pydantic-settings==2.0.3
pytest==7.4.3
httpx==0.25.1
pytestcov==4.1.0
